app/requirements.txt
flask
flask-wtf
requests
python-firebase
